cart button styles finished
