+++ libs/shared/ui/src/components/navigation/NavLogo.tsx
@@ -4,7 +4,7 @@ import { SofaIcon } from '@sas-mrts/common'
 
 function NavLogo() {
   return (
-    <Button asChild className="p-2">
+    <Button asChild className="p-2 cursor-pointer">
       <SofaIcon className="h-12 w-12" />
     </Button>
   )
