+++ libs/shared/ui/src/components/navigation/actions/NavActions.tsx
@@ -1,0 +1,13 @@
+import CartButton from './CartButton'
+import ThemeToggle from './ThemeToggle'
+
+function NavActions() {
+  return (
+    <div className="flex gap-4 m-0 p-0">
+      <ThemeToggle />
+      <CartButton />
+    </div>
+  )
+}
+
+export default NavActions
